Relocate Test Utility Classes
