+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/GeneratedEnumTests.java
@@ -22,15 +22,16 @@ import com.google.gson.JsonParser;
 import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
-import io.github.chrimle.o2jrm.GeneratedSource;
-import io.github.chrimle.o2jrm.configs.ConfigOption;
-import io.github.chrimle.o2jrm.configs.Library;
-import io.github.chrimle.o2jrm.models.GeneratedField;
 import io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne;
 import io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree;
 import io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo;
 import io.github.chrimle.o2jrm.test.interfaces.TestInterfaceOne;
 import io.github.chrimle.o2jrm.test.interfaces.TestInterfaceTwo;
+import io.github.chrimle.o2jrm.test.meta.configs.ConfigOption;
+import io.github.chrimle.o2jrm.test.meta.configs.Library;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedField;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedSource;
+import io.github.chrimle.o2jrm.test.utils.CustomAssertions;
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests;
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests.AdditionalEnumTypeAnnotationsTests;
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests.UseEnumCaseInsensitiveTests;
@@ -38,7 +39,6 @@ import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.OpenAPITests.SchemaTests
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.OpenAPITests.SchemaTests.DeprecatedTests;
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.OpenAPITests.SchemaTests.EnumTests;
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.OpenAPITests.SchemaTests.EnumTests.ConstantsTests;
-import io.github.chrimle.o2jrm.utils.CustomAssertions;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.Serializable;
