+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/GeneratedRecordTests.java
@@ -20,13 +20,16 @@ import com.google.gson.*;
 import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
-import io.github.chrimle.o2jrm.GeneratedSource;
-import io.github.chrimle.o2jrm.configs.ConfigOption;
-import io.github.chrimle.o2jrm.configs.Library;
-import io.github.chrimle.o2jrm.models.GeneratedField;
 import io.github.chrimle.o2jrm.test.annotations.*;
 import io.github.chrimle.o2jrm.test.interfaces.TestInterfaceOne;
 import io.github.chrimle.o2jrm.test.interfaces.TestInterfaceTwo;
+import io.github.chrimle.o2jrm.test.meta.configs.ConfigOption;
+import io.github.chrimle.o2jrm.test.meta.configs.Library;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedField;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedSource;
+import io.github.chrimle.o2jrm.test.utils.AssertionUtils;
+import io.github.chrimle.o2jrm.test.utils.CustomAssertions;
+import io.github.chrimle.o2jrm.test.utils.GeneratedRecordTestUtils;
 import io.github.chrimle.o2jrm.tests.GeneratedRecordTests.GeneratorConfigurationTests.ConfigOptionsTests;
 import io.github.chrimle.o2jrm.tests.GeneratedRecordTests.GeneratorConfigurationTests.ConfigOptionsTests.AdditionalModelTypeAnnotationsTests;
 import io.github.chrimle.o2jrm.tests.GeneratedRecordTests.GeneratorConfigurationTests.ConfigOptionsTests.GenerateBuildersTests;
@@ -39,9 +42,6 @@ import io.github.chrimle.o2jrm.tests.GeneratedRecordTests.OpenAPITests.SchemaTes
 import io.github.chrimle.o2jrm.tests.GeneratedRecordTests.OpenAPITests.SchemaTests.PropertiesTests.PropertyTests.XFieldExtraAnnotationTests;
 import io.github.chrimle.o2jrm.tests.GeneratedRecordTests.OpenAPITests.SchemaTests.TypeTests;
 import io.github.chrimle.o2jrm.tests.GeneratedRecordTests.OpenAPITests.SchemaTests.XClassExtraAnnotationTests;
-import io.github.chrimle.o2jrm.utils.AssertionUtils;
-import io.github.chrimle.o2jrm.utils.CustomAssertions;
-import io.github.chrimle.o2jrm.utils.GeneratedRecordTestUtils;
 import java.io.*;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Constructor;
