+++ test-java/src/test/java/io/github/chrimle/o2jrm/models/GeneratedRecordImpl.java
@@ -15,13 +15,16 @@
  */
 package io.github.chrimle.o2jrm.models;
 
-import io.github.chrimle.o2jrm.PluginExecution;
 import io.github.chrimle.o2jrm.test.annotations.TestExtraAnnotation;
 import io.github.chrimle.o2jrm.test.annotations.TestExtraAnnotationTwo;
 import io.github.chrimle.o2jrm.test.annotations.TestFieldExtraAnnotationOne;
 import io.github.chrimle.o2jrm.test.annotations.TestFieldExtraAnnotationTwo;
 import io.github.chrimle.o2jrm.test.interfaces.TestInterfaceOne;
 import io.github.chrimle.o2jrm.test.interfaces.TestInterfaceTwo;
+import io.github.chrimle.o2jrm.test.meta.configs.PluginExecution;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedClass;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedField;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedRecord;
 import java.lang.annotation.Annotation;
 import java.math.BigDecimal;
 import java.util.List;
