+++ test-utils/src/main/java/io/github/chrimle/o2jrm/test/meta/configs/Library.java
@@ -14,16 +14,20 @@
  * limitations under the License.
  */
 
-package io.github.chrimle.o2jrm.models;
+package io.github.chrimle.o2jrm.test.meta.configs;
 
-public enum BeanValidationAnnotation {
-  VALID,
-  NOT_NULL,
-  PATTERN,
-  SIZE,
-  MIN,
-  MAX,
-  DECIMAL_MIN,
-  DECIMAL_MAX,
-  EMAIL
+public enum Library {
+  OKHTTP_GSON("okhttp-gson"),
+  SPRING_BOOT("spring-boot"),
+  WEBCLIENT("webclient");
+
+  private final String id;
+
+  Library(final String id) {
+    this.id = id;
+  }
+
+  public String getId() {
+    return id;
+  }
 }
