+++ test-utils/src/main/java/io/github/chrimle/o2jrm/test/meta/models/GeneratedClass.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.chrimle.o2jrm.models;
+package io.github.chrimle.o2jrm.test.meta.models;
 
-import io.github.chrimle.o2jrm.PluginExecution;
+import io.github.chrimle.o2jrm.test.meta.configs.PluginExecution;
 import java.lang.annotation.Annotation;
 import java.util.List;
