+++ test-utils/src/main/java/io/github/chrimle/o2jrm/test/meta/models/GeneratedEnum.java
@@ -13,10 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.chrimle.o2jrm.models;
+package io.github.chrimle.o2jrm.test.meta.models;
 
-/** Expected {@code record} class to be generated from the OpenAPI spec. */
-public interface GeneratedRecord extends GeneratedClass {
+import java.lang.annotation.Annotation;
+import java.util.List;
+
+/** Expected {@code enum} class to be generated from the OpenAPI spec. */
+public interface GeneratedEnum extends GeneratedClass {
 
   /**
    * {@inheritDoc}
@@ -25,7 +28,7 @@ public interface GeneratedRecord extends GeneratedClass {
    */
   @Override
   default boolean isEnum() {
-    return false;
+    return true;
   }
 
   /**
@@ -35,6 +38,26 @@ public interface GeneratedRecord extends GeneratedClass {
    */
   @Override
   default boolean isInnerEnum() {
+    return getSimpleClassName().contains("$");
+  }
+
+  /**
+   * {@inheritDoc}
+   *
+   * @return whether the class has extra annotations.
+   */
+  @Override
+  default boolean hasExtraAnnotations() {
     return false;
+  }
+
+  /**
+   * {@inheritDoc}
+   *
+   * @return the collection of annotations.
+   */
+  @Override
+  default List<Class<? extends Annotation>> getExtraAnnotations() {
+    return List.of();
   }
 }
