+++ test-utils/src/main/java/io/github/chrimle/o2jrm/test/utils/CustomUtilityAssertions.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.chrimle.o2jrm.utils;
+package io.github.chrimle.o2jrm.test.utils;
 
 import java.util.function.Supplier;
 import org.junit.jupiter.api.Assertions;
