+++ test-utils/src/main/java/io/github/chrimle/o2jrm/test/utils/GeneratedRecordTestUtils.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.chrimle.o2jrm.utils;
+package io.github.chrimle.o2jrm.test.utils;
 
-import io.github.chrimle.o2jrm.GeneratedSource;
-import io.github.chrimle.o2jrm.models.GeneratedField;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedField;
+import io.github.chrimle.o2jrm.test.meta.models.GeneratedSource;
 import java.lang.reflect.Constructor;
 import java.math.BigDecimal;
 import java.util.ArrayList;
